# doc/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Home Library Service",
    "description": "Home music library service",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "schemas": {
...
      "Album": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Innuendo"
          },
          "year": {
            "type": "integer",
            "example": 1991
          },
          "artistId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Access token is missing or invalid"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/albums/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "tags": [
          "Album"
        ],
        "summary": "Get single album by id",
        "description": "Gets single album by id",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Album"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. albumId is invalid (not uuid)"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Album was not found."
          }
        }
      },
      "put": {
        "tags": [
          "Album"
        ],
        "summary": "Update album information",
        "description": "Update library album information by UUID",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "example",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "year": {
                    "type": "integer"
                  },
                  "artistId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  }
                },
                "required": [
                  "name",
                  "year"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The album has been updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Album"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. albumId is invalid (not uuid)"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Album was not found."
          }
        }
      },
      "delete": {
        "tags": [
          "Album"
        ],
        "summary": "Delete album",
        "description": "Delete album from library",
        "responses": {
          "204": {
            "description": "Deleted successfully"
          },
          "400": {
            "description": "Bad request. albumId is invalid (not uuid)"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Album was not found."
          }
        }
      }
    },
...
    "/favs/album/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "tags": [
          "Favorites"
        ],
        "summary": "Add album to the favorites",
        "description": "Add album to the favorites",
        "responses": {
          "201": {
            "description": "Added successfully"
          },
          "400": {
            "description": "Bad. albumId is invalid (not uuid)"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "422": {
            "description": "Album with id doesn't exist."
          }
        }
      },
      "delete": {
        "tags": [
          "Favorites"
        ],
        "summary": "Delete album from favorites",
        "description": "Delete album from favorites",
        "responses": {
          "204": {
            "description": "Deleted successfully"
          },
          "400": {
            "description": "Bad. albumId is invalid (not uuid)"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Album was not found."
          }
        }
      }
    },
...
  }
}